+++ src/lib/validate.h
@@ -42,6 +42,7 @@ bool vcb_ismpdtag(sds data);
 bool vcb_ismpdtag_or_any(sds data);
 bool vcb_ismpdstickertype(sds data);
 bool vcb_ismpdsort(sds data);
+bool vcb_isplaylistsort(sds data);
 bool vcb_iswebradiosort(sds data);
 bool vcb_issearchexpression(sds data);