+++ src/mympd_api/mympd_api_handler.c
@@ -1341,10 +1341,13 @@ void mympd_api_handler(struct t_mympd_state *mympd_state, struct t_partition_sta
                 json_get_uint(request->data, "$.params.limit", MPD_RESULTS_MIN, MPD_RESULTS_MAX, &uint_buf2, &parse_error) == true &&
                 json_get_string(request->data, "$.params.searchstr", 0, NAME_LEN_MAX, &sds_buf1, vcb_isname, &parse_error) == true &&
                 json_get_uint(request->data, "$.params.type", 0, 2, &uint_buf3, &parse_error) == true &&
+                json_get_string(request->data, "$.params.sort", 0, NAME_LEN_MAX, &sds_buf2, vcb_ismpdsort, &parse_error) == true &&
+                json_get_bool(request->data, "$.params.sortdesc", &bool_buf1, &parse_error) == true &&
                 json_get_fields(request->data, "$.params.fields", &tagcols, FIELDS_MAX, &parse_error) == true)
             {
+                enum playlist_sort_types sort = playlist_parse_sort(sds_buf2);
                 response->data = mympd_api_playlist_list(partition_state, mympd_state->stickerdb, response->data, request->cmd_id,
-                    uint_buf1, uint_buf2, sds_buf1, uint_buf3, &tagcols);
+                    uint_buf1, uint_buf2, sds_buf1, uint_buf3, sort, bool_buf1, &tagcols);
             }
             break;
         }